+++ internal/net/conn.go
@@ -5,6 +5,7 @@ import (
 	"net"
 	"time"
 
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/errors"
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
@@ -21,7 +22,7 @@ func New(ctx context.Context, address string, t trace.Driver) (_ net.Conn, err e
 	}()
 	cc, err := (&net.Dialer{}).DialContext(ctx, "tcp", address)
 	if err != nil {
-		return nil, err
+		return nil, errors.Errorf(0, "dial failed: %w", err)
 	}
 	return &conn{
 		address: address,
@@ -35,7 +36,11 @@ func (c *conn) Read(b []byte) (n int, err error) {
 	defer func() {
 		onDone(n, err)
 	}()
-	return c.cc.Read(b)
+	n, err = c.cc.Read(b)
+	if err != nil {
+		return n, errors.Errorf(0, "read failed: %w", err)
+	}
+	return n, nil
 }
 
 func (c *conn) Write(b []byte) (n int, err error) {
@@ -43,7 +48,11 @@ func (c *conn) Write(b []byte) (n int, err error) {
 	defer func() {
 		onDone(n, err)
 	}()
-	return c.cc.Write(b)
+	n, err = c.cc.Write(b)
+	if err != nil {
+		return n, errors.Errorf(0, "write failed: %w", err)
+	}
+	return n, nil
 }
 
 func (c *conn) Close() (err error) {
@@ -51,7 +60,11 @@ func (c *conn) Close() (err error) {
 	defer func() {
 		onDone(err)
 	}()
-	return c.cc.Close()
+	err = c.cc.Close()
+	if err != nil {
+		return errors.Errorf(0, "close failed: %w", err)
+	}
+	return nil
 }
 
 func (c *conn) LocalAddr() net.Addr {
@@ -62,7 +75,7 @@ func (c *conn) RemoteAddr() net.Addr {
 	return c.cc.RemoteAddr()
 }
 
-func (c *conn) SetDeadline(t time.Time) error {
+func (c *conn) SetDeadline(t time.Time) (err error) {
 	return c.cc.SetDeadline(t)
 }