+++ internal/table/client.go
@@ -40,7 +40,7 @@ var (
 )
 
 func errNoProgress(attempts int) error {
-	return fmt.Errorf("ydb: table: no progress: %w (%d attempts)", ErrNoProgress, attempts)
+	return errors.Errorf(1, "ydb: table: no progress: %w (%d attempts)", ErrNoProgress, attempts)
 }
 
 // SessionBuilder is the interface that holds logic of creating sessions.
@@ -200,28 +200,39 @@ func (c *client) createSession(ctx context.Context) (s Session, err error) {
 			panic("ydb: abnormal result of session build")
 		}
 
-		c.mu.Lock()
-		c.createInProgress--
 		if s != nil {
 			s.OnClose(func(ctx context.Context) {
 				c.mu.Lock()
 				defer c.mu.Unlock()
+
 				info, has := c.index[s]
 				if !has {
 					return
 				}
+
 				delete(c.index, s)
+
+				trace.TableOnPoolStateChange(c.config.Trace(), len(c.index), "remove")
 
 				if c.closed {
 					return
 				}
 
 				c.notify(nil)
+
 				if info.idle != nil {
 					panic("ydb: table: session closed while still in idle client")
 				}
 			})
-			c.index[s] = sessionInfo{}
+		}
+
+		c.mu.Lock()
+		{
+			c.createInProgress--
+			if s != nil {
+				c.index[s] = sessionInfo{}
+				trace.TableOnPoolStateChange(c.config.Trace(), len(c.index), "append")
+			}
 		}
 		c.mu.Unlock()
 
@@ -283,6 +294,9 @@ func (c *client) Get(ctx context.Context) (s Session, err error) {
 		}
 		// got session or err is not recoverable
 		if s != nil || !isCreateSessionErrorRetriable(err) {
+			if err != nil {
+				err = errors.Errorf(1, "create session failed: %w", err)
+			}
 			return s, err
 		}
 
@@ -293,9 +307,15 @@ func (c *client) Get(ctx context.Context) (s Session, err error) {
 		// be fair here and not to lock more goroutines than we could ship
 		// session to.
 		s, err = c.waitFromCh(ctx, t)
+		if err != nil {
+			err = errors.Errorf(1, "wait from channel failed: %w", err)
+		}
 	}
 	if s == nil && err == nil {
 		err = errNoProgress(i)
+	}
+	if err != nil {
+		err = errors.Errorf(1, "get failed: %w", err)
 	}
 	return s, err
 }
@@ -397,95 +417,6 @@ func (c *client) Put(ctx context.Context, s Session) (err error) {
 	return err
 }
 
-// Take removes session from the client and ensures that s will not be returned
-// by other Take() or Get() calls.
-//
-// The intended way of Take() use is to create session by calling Create() and
-// Put() it later to prepare KeepAlive tracking when session is idle. When
-// session becomes active, one should call Take() to stop KeepAlive tracking
-// (simultaneous use of session is prohibited).
-//
-// After session returned to the client by calling PutBusy() it can not be taken
-// by Take() any more. That is, semantically PutBusy() is the same as session's
-// Close().
-//
-// It is assumed that Take() callers never call Get() method.
-func (c *client) Take(ctx context.Context, s Session) (took bool, err error) {
-	onWait := trace.TableOnPoolTake(c.config.Trace().Compose(trace.ContextTable(ctx)), &ctx, s)
-	var onDone func(took bool, _ error)
-	defer func() {
-		if onDone == nil {
-			onDone = onWait()
-		}
-		onDone(took, err)
-	}()
-
-	if c.isClosed() {
-		return false, ErrSessionPoolClosed
-	}
-
-	var has bool
-	c.mu.Lock()
-	for has, took = c.takeIdle(s); has && !took && c.touching; has, took = c.takeIdle(s) {
-		cond := c.touchCond()
-		c.mu.Unlock()
-		onDone = onWait()
-
-		// Keepalive processing takes place right now.
-		// Try to await touched session before creation of new one.
-		select {
-		case <-cond:
-
-		case <-ctx.Done():
-			return false, ctx.Err()
-		}
-
-		c.mu.Lock()
-	}
-	c.mu.Unlock()
-
-	if !has {
-		err = ErrSessionUnknown
-	}
-
-	return took, err
-}
-
-// Create creates new session and returns it.
-// The intended way of Create() usage relates to Take() method.
-func (c *client) Create(ctx context.Context) (s Session, err error) {
-	const maxAttempts = 10
-	i := 0
-	for ; i < maxAttempts; i++ {
-		c.mu.Lock()
-		// NOTE: here is a race condition with keeper() running.
-		// session could be deleted by some reason after we released the mutex.
-		// We are not dealing with this because even if session is not deleted
-		// by keeper() it could be staled on the server and the same user
-		// experience will appear.
-		s, _ = c.peekFirstIdle()
-		c.mu.Unlock()
-
-		if s == nil {
-			return c.createSession(ctx)
-		}
-
-		took, e := c.Take(ctx, s)
-		if e == nil && !took || errors.Is(e, ErrSessionUnknown) {
-			// session was marked for deletion or deleted by keeper() - race happen - retry
-			s = nil
-			continue
-		}
-		if e != nil {
-			return nil, e
-		}
-
-		return s, nil
-	}
-
-	return nil, errNoProgress(i)
-}
-
 // Close deletes all stored sessions inside client.
 // It also stops all underlying timers and goroutines.
 // It returns first error occurred during stale sessions' deletion.
@@ -576,29 +507,6 @@ func (c *client) DoTx(ctx context.Context, op table.TxOperation, opts ...table.O
 		withOptions(opts...),
 		withTrace(c.config.Trace()),
 	)
-}
-
-func (c *client) Stats() poolStats {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	idleCount, waitQCount, indexCount := 0, 0, 0
-	if c.idle != nil {
-		idleCount = c.idle.Len()
-	}
-	if c.waitq != nil {
-		waitQCount = c.waitq.Len()
-	}
-	if c.index != nil {
-		indexCount = len(c.index)
-	}
-	return poolStats{
-		Idle:             idleCount,
-		Index:            indexCount,
-		WaitQ:            waitQCount,
-		CreateInProgress: c.createInProgress,
-		MinSize:          c.config.KeepAliveMinSize(),
-		MaxSize:          c.limit,
-	}
 }
 
 func (c *client) keeper() {
@@ -887,10 +795,14 @@ func (c *client) closeSession(ctx context.Context, s Session) error {
 	return nil
 }
 
-func (c *client) keepAliveSession(ctx context.Context, s Session) error {
+func (c *client) keepAliveSession(ctx context.Context, s Session) (err error) {
 	ctx, cancel := context.WithTimeout(ctx, c.config.KeepAliveTimeout())
 	defer cancel()
-	return s.KeepAlive(ctx)
+	err = s.KeepAlive(ctx)
+	if err != nil {
+		return errors.Errorf(0, "keep-alive session failed: %w", err)
+	}
+	return nil
 }
 
 // p.mu must be held.
@@ -904,29 +816,6 @@ func (c *client) removeIdle(s Session) sessionInfo {
 	info.idle = nil
 	c.index[s] = info
 	return info
-}
-
-// Removes session from idle client and resets keepAliveCount for it not
-// to die in keeper when it will be returned
-// to be used only in outgoing functions that make session busy.
-// p.mu must be held.
-func (c *client) takeIdle(s Session) (has, took bool) {
-	var info sessionInfo
-	info, has = c.index[s]
-	if !has {
-		// Could not be strict here and panic – session may become deleted by
-		// keeper().
-		return
-	}
-	if info.idle == nil {
-		// session s is not idle.
-		return
-	}
-	took = true
-	info = c.removeIdle(s)
-	info.keepAliveCount = 0
-	c.index[s] = info
-	return
 }
 
 // p.mu must be held.
@@ -1001,13 +890,4 @@ type sessionInfo struct {
 func panicLocked(mu sync.Locker, message string) {
 	mu.Unlock()
 	panic(message)
-}
-
-type poolStats struct {
-	Idle             int
-	Index            int
-	WaitQ            int
-	MinSize          int
-	MaxSize          int
-	CreateInProgress int
 }